+++ lib/pages/groups.dart
@@ -5,6 +5,7 @@ import 'package:poll_and_play/pages/group.dart';
 import 'package:poll_and_play/pages/page.dart' as page;
 import 'package:poll_and_play/providers/groups.dart';
 import 'package:poll_and_play/providers/state.dart';
+import 'package:poll_and_play/ui/cross_platform_refresh.dart';
 import 'package:poll_and_play/ui/dialog_button.dart';
 import 'package:poll_and_play/ui/text_input.dart';
 import 'package:poll_play_proto_gen/public.dart';
@@ -40,21 +41,11 @@ class GroupsPage extends StatelessWidget implements page.Page {
   Widget build(BuildContext context) {
     GroupsProvider provider = Provider.of<GroupsProvider>(context);
 
-    return ScrollConfiguration(
-      behavior: ScrollConfiguration.of(context).copyWith(
-        physics: const BouncingScrollPhysics(),
-        dragDevices: {
-          PointerDeviceKind.touch,
-          PointerDeviceKind.mouse,
-          PointerDeviceKind.trackpad
-        },
-      ),
-      child: RefreshIndicator(
-        onRefresh: provider.refresh,
-        child: ListView(
-          padding: const EdgeInsets.all(8),
-          children: _generateGroups(context),
-        ),
+    return CrossPlatformRefreshIndicator(
+      onRefresh: provider.refresh,
+      child: ListView(
+        padding: const EdgeInsets.all(8),
+        children: _generateGroups(context),
       ),
     );
   }